refactor(pipeline/network): extract Debug from NetworkDiscovery
Phase 4 commit 1/9 of NetworkDiscovery split (master plan
pipeline-decoupling-master-2026-05-28).

- Carve Debug (dumpCounter, IDumpArgs, dumpResponseBody) from
  NetworkDiscovery.ts (1811 -> 1746 LoC).
- New file: src/Scrapers/Pipeline/Mediator/Network/Debug/NetworkDump.ts
  (~81 LoC).
- NetworkDiscovery.ts now imports dumpResponseBody from the new
  sub-module; drops node:fs / node:path / TraceConfig / ActiveState
  imports that became unused.
- No behavior change; no test BODY change; public API unchanged.

Master plan: pipeline-decoupling-master-2026-05-28 / phase-4
